+++ CHANGELOG.md
@@ -5,3 +5,7 @@
 ## v0.0.5:
 1. add rest of request methods: "head", "options", "put", "patch", "delete"
 2. change the second param of post(request alias) as "data"(data in config before) and the third param is "config"
+
+## v0.0.6: (TO BE SUPPORT)
+1. fix ways to get response headers
+2. auto patch retry times for cache's lead request